modificacion de estructura
