+++ mi_app/forms/__init__.py
@@ -1,0 +1,6 @@
+from .document_type_form import DocumentTypeForm
+from .payment_type_form import PaymentTypeForm
+from .predetermined_price_form import PredeterminedPriceForm
+from .patient_form import PatientForm
+from .history_form import HistoryForm
+from .appointment_form import AppointmentForm
