+++ mi_app/models/__init__.py
@@ -1,0 +1,6 @@
+from .document_type import DocumentType
+from .patient import Patient
+from .history import History
+from .payment_type import PaymentType
+from .predetermined_price import PredeterminedPrice
+from .appointment import Appointment
