+++ mi_app/models/predetermined_price.py
@@ -1,0 +1,17 @@
+from django.db import models
+
+class PredeterminedPrice(models.Model):
+    name = models.CharField(
+        max_length=255,
+        error_messages={'max_length': 'El nombre no debe superar los 255 caracteres.'}
+    )
+    price = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    def __str__(self):
+        return f"{self.name} - {self.price}"
+
+    class Meta:
+        db_table = "predetermined_prices"
