--- mi_app/templates/document_types/list.html
@@ -1,8 +1,0 @@
-<h1>Tipos de Documento</h1>
-<ul>
-  {% for doc_type in document_types %}
-    <li>{{ doc_type.name }}</li>
-  {% empty %}
-    <li>No hay tipos de documento.</li>
-  {% endfor %}
-</ul>
