+++ mi_app/views/__init__.py
@@ -1,0 +1,6 @@
+from .document_type_views import *
+from .patient_views import *
+from .history_views import *
+from .payment_type_views import *
+from .predetermined_price_views import *
+from .appointment_views import *
